requirements-full.txt
-r requirements.txt

absl-py==2.3.1
astunparse==1.6.3
cachetools==5.5.2
contourpy==1.3.2
cycler==0.12.1
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.0
fsspec==2025.7.0
gast==0.4.0
google-auth==2.40.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
imageio==2.37.0
joblib==1.5.1
keras==3.11.2
Keras-Preprocessing==1.1.2
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
matplotlib==3.10.5
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
namex==0.1.0
networkx==3.4.2
oauthlib==3.3.1
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
pyparsing==3.2.3
requests-oauthlib==2.0.0
rich==14.1.0
rsa==4.9.1
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.19.0
tensorflow-addons==0.22.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typeguard==2.13.3
typing-inspection==0.4.1
ultralytics==8.3.179
ultralytics-thop==2.0.15
Werkzeug==3.1.3
wrapt==1.17.2
